requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
psycopg2-binary==2.9.3
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
